src/cloud_run/requirements.txt
fastapi[standard]
celery
flower
redis
google-cloud-storage
sendgrid
openai
moviepy
elevenlabs